+++ angular/config/theme.js
@@ -1,7 +1,7 @@
 (function(){
 	"use strict";
 
-	angular.module('app.config').config( function($mdThemingProvider) {
+	angular.module('app.config').config(function($mdThemingProvider) {
 		/* For more info, visit https://material.angularjs.org/#/Theming/01_introduction */
 		$mdThemingProvider.theme('default')
 		.primaryPalette('indigo')