+++ gulpfile.js
@@ -1,6 +1,7 @@
 var elixir = require('laravel-elixir');
 require('./tasks/angular.task.js');
 require('./tasks/bower.task.js');
+require('laravel-elixir-livereload');
 
 process.env.DISABLE_NOTIFIER = true;
 
@@ -22,5 +23,12 @@ elixir(function(mix){
 		.less('./angular/**/*.less', 'public/css')
 		.copy('./angular/app/**/*.html', 'public/views/app/')
 		.copy('./angular/directives/**/*.html', 'public/views/directives/')
-		.copy('./angular/dialogs/**/*.html', 'public/views/dialogs/');
+		.copy('./angular/dialogs/**/*.html', 'public/views/dialogs/')
+		.livereload([
+			'public/js/vendor.js',
+			'public/js/app.js',
+			'public/css/vendor.css',
+			'public/css/app.css',
+			'public/views/**/*.html'
+		]);
 });