+++ resources/views/index.blade.php
@@ -21,5 +21,11 @@
 <script src="/js/app.js"></script>
 <script src="/js/prism.js"></script>
 
+{{--livereload--}}
+@if ( Config::get('app.debug') )
+    <script type="text/javascript">
+        document.write('<script src="//localhost:35729/livereload.js?snipver=1" type="text/javascript"><\/script>')
+    </script>
+@endif
 </body>
 </html>